Release v2.8.12: Fix all compilation warnings and update version
- Fix UTF-8 encoding warnings (C4819) by converting files to ANSI encoding
- Add compiler flags /wd4819 and -Xcompiler=/wd4819 to suppress encoding warnings
- Fix type conversion warning in GPUEngine.cu with static_cast<int>()
- Replace deprecated network APIs: inet_ntoa -> inet_ntop, gethostbyname -> getaddrinfo
- Add ws2tcpip.h header for modern Windows network functions
- Update version number from v2.2 to v2.8.12 in Constants.h
- Add encoding fix PowerShell script for future maintenance
- Achieve zero-warning compilation suitable for production deployment
- Maintain all existing functionality and performance characteristics

+++ GPU/GPUEngine.cu
@@ -291,7 +291,7 @@ GPUEngine::~GPUEngine() {
 
 
 int GPUEngine::GetMemory() {
-  return kangarooSize + outputSize + jumpSize;
+  return static_cast<int>(kangarooSize + outputSize + jumpSize);
 }
